move EnergyWithPercentageWithCO2ePerMWh to common
