+++ src/climatevision/generator/business2018/energy_source/dataclasses.py
@@ -3,7 +3,7 @@
 from dataclasses import dataclass
 
 from ...utils import MILLION
-from ...common.energy import EnergyWithPercentage
+from ...common.energy_with_co2e_per_mwh import EnergyWithPercentageWithCO2ePerMWh
 
 
 @dataclass(kw_only=True)
@@ -13,19 +13,6 @@ class Vars5:
     cost_fuel: float = None  # type: ignore
     energy: float = None  # type: ignore
     pct_energy: float = None  # type: ignore
-
-
-@dataclass(kw_only=True)
-class EnergyWithPercentageWithCO2ePerMWh(EnergyWithPercentage):
-    CO2e_combustion_based: float = 0
-    CO2e_combustion_based_per_MWh: float
-    CO2e_total: float = 0
-
-    def __post_init__(self, total_energy: float):
-        self.CO2e_combustion_based = self.energy * self.CO2e_combustion_based_per_MWh
-        self.CO2e_total = self.CO2e_combustion_based
-
-        EnergyWithPercentage.__post_init__(self, total_energy)
 
 
 @dataclass(kw_only=True)
