+++ src/climatevision/generator/common/energy_with_co2e_per_mwh.py
@@ -5,6 +5,7 @@ from dataclasses import dataclass
 from ..utils import div
 
 from .co2_equivalent_emission import CO2eEmission
+from .energy import EnergyWithPercentage
 from .energy_with_co2e import EnergyWithCO2e
 
 
@@ -39,3 +40,16 @@ class EnergyWithCO2ePerMWh(EnergyWithCO2e):
             CO2e_production_based_per_MWh=CO2e_production_based_per_MWh,
             CO2e_total=CO2e_total,
         )
+
+
+@dataclass(kw_only=True)
+class EnergyWithPercentageWithCO2ePerMWh(EnergyWithPercentage):
+    CO2e_combustion_based: float = 0
+    CO2e_combustion_based_per_MWh: float
+    CO2e_total: float = 0
+
+    def __post_init__(self, total_energy: float):
+        self.CO2e_combustion_based = self.energy * self.CO2e_combustion_based_per_MWh
+        self.CO2e_total = self.CO2e_combustion_based
+
+        EnergyWithPercentage.__post_init__(self, total_energy)
